Post and delete methods

+++ src/main/java/tc3005b224/amazonconnectinsights/service/InsightsService.java
@@ -4,6 +4,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tc3005b224.amazonconnectinsights.dto.insights.InsightsDTO;
+import tc3005b224.amazonconnectinsights.models_sql.Category;
 import tc3005b224.amazonconnectinsights.models_sql.Insight;
 import tc3005b224.amazonconnectinsights.repository.InsightRepository;
 
@@ -35,10 +36,22 @@ public class InsightsService {
         return null;
     }
 
-    public void deleteInsight(Short id) {
-        insightRepository.deleteById(id);
+    public boolean deleteInsight(Short id) {
+        if (insightRepository.existsById(id)) {
+            insightRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
+    public InsightsDTO createInsight(InsightsDTO insightDTO) {
+        Insight insight = toInsight(insightDTO);
+        insight = insightRepository.save(insight);
+        return toInsightDTO(insight);
+    }
+
+
+    
     private InsightsDTO toInsightDTO(Insight insight) {
         InsightsDTO dto = new InsightsDTO();
         dto.setIdentifier(insight.getIdentifier());
@@ -51,10 +64,18 @@ public class InsightsService {
         return dto;
     }
 
-    private Insight toInsight(InsightDTO dto) {
+    private Insight toInsight(InsightsDTO dto) {
         Insight insight = new Insight();
-
+        
+        insight.setIdentifier(dto.getIdentifier());
+        insight.setCategory(dto.getCategoryIdentifier());
+        insight.setDenomination(dto.getDenomination());
+        insight.setDescription(dto.getDescription());
+        insight.setDateRegistered(dto.getDateRegistered());
+        insight.setDateUpdated(dto.getDateUpdated());
+        insight.setActive(dto.getActive());
         return insight;
     }
+
     
 }
